# web/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Anchor": {
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "points": {
            "anyOf": [
              {
                "items": {
                  "maxItems": 3,
                  "minItems": 3,
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "position": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Position",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "position"
        ],
        "title": "Anchor",
        "type": "object"
      },
      "AnchorIn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "points": {
            "items": {
              "maxItems": 3,
              "minItems": 3,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Points",
            "type": "array"
          },
          "position": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Position",
            "type": "array"
          }
        },
        "required": [
          "name",
          "position"
        ],
        "title": "AnchorIn",
        "type": "object"
      },
      "AnchorUpdate": {
        "properties": {
          "confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "points": {
            "anyOf": [
              {
                "items": {
                  "maxItems": 3,
                  "minItems": 3,
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "position": {
            "anyOf": [
              {
                "maxItems": 3,
                "minItems": 3,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "title": "AnchorUpdate",
        "type": "object"
      },
...
      "Box2": {
        "properties": {
          "max": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Max",
            "type": "array"
          },
          "min": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Min",
            "type": "array"
          }
        },
        "required": [
          "min",
          "max"
        ],
        "title": "Box2",
        "type": "object"
      },
...
      "Furniture": {
        "properties": {
          "box": {
            "anyOf": [
              {
                "maxItems": 4,
                "minItems": 4,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Box"
          },
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "height": {
            "title": "Height",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "polygon": {
            "anyOf": [
              {
                "items": {
                  "maxItems": 2,
                  "minItems": 2,
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon"
          },
          "z0": {
            "default": 0.0,
            "title": "Z0",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "height"
        ],
        "title": "Furniture",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Home": {
        "properties": {
          "anchors": {
            "items": {
              "$ref": "#/components/schemas/Anchor"
            },
            "title": "Anchors",
            "type": "array"
          },
          "beams": {
            "title": "Beams",
            "type": "number"
          },
          "ceiling": {
            "title": "Ceiling",
            "type": "number"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/Box2"
            },
            "title": "Columns",
            "type": "array"
          },
          "furniture": {
            "items": {
              "$ref": "#/components/schemas/Furniture"
            },
            "title": "Furniture",
            "type": "array"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "northOffsetDeg": {
            "title": "Northoffsetdeg",
            "type": "number"
          },
          "outdoor": {
            "$ref": "#/components/schemas/Outdoor"
          },
          "outline": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Outline",
            "type": "array"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/Room"
            },
            "title": "Rooms",
            "type": "array"
          },
          "size": {
            "$ref": "#/components/schemas/HomeSize"
          },
          "subZones": {
            "items": {
              "$ref": "#/components/schemas/SubZone"
            },
            "title": "Subzones",
            "type": "array"
          },
          "wallCutHeight": {
            "title": "Wallcutheight",
            "type": "number"
          },
          "walls": {
            "items": {
              "$ref": "#/components/schemas/Wall"
            },
            "title": "Walls",
            "type": "array"
          }
        },
        "required": [
          "outline",
          "rooms",
          "subZones",
          "walls",
          "columns",
          "furniture",
          "anchors",
          "ceiling",
          "beams",
          "wallCutHeight",
          "size",
          "northOffsetDeg"
        ],
        "title": "Home",
        "type": "object"
      },
      "HomeSettings": {
        "description": "PUT /home: only what's sent changes.",
        "properties": {
          "beams": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beams"
          },
          "ceiling": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ceiling"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "northOffsetDeg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Northoffsetdeg"
          }
        },
        "title": "HomeSettings",
        "type": "object"
      },
      "HomeSize": {
        "properties": {
          "eastWest": {
            "title": "Eastwest",
            "type": "number"
          },
          "northSouth": {
            "title": "Northsouth",
            "type": "number"
          }
        },
        "required": [
          "eastWest",
          "northSouth"
        ],
        "title": "HomeSize",
        "type": "object"
      },
...
      "Location": {
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "lat",
          "lon"
        ],
        "title": "Location",
        "type": "object"
      },
...
      "Outdoor": {
        "properties": {
          "balcony": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Balcony",
            "type": "array"
          },
          "balconyOffRoom": {
            "default": "",
            "title": "Balconyoffroom",
            "type": "string"
          },
          "courtyard": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Courtyard",
            "type": "array"
          },
          "courtyardOpensTo": {
            "default": "",
            "title": "Courtyardopensto",
            "type": "string"
          }
        },
        "title": "Outdoor",
        "type": "object"
      },
...
      "Placement": {
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "confirmedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmedat"
          },
          "ledOrder": {
            "title": "Ledorder",
            "type": "string"
          },
          "shape": {
            "discriminator": {
              "mapping": {
                "bent-line": "#/components/schemas/BentLineShape",
                "cylinder": "#/components/schemas/CylinderShape",
                "grid": "#/components/schemas/GridShape",
                "line": "#/components/schemas/LineShape",
                "point": "#/components/schemas/PointShape"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PointShape"
              },
              {
                "$ref": "#/components/schemas/LineShape"
              },
              {
                "$ref": "#/components/schemas/BentLineShape"
              },
              {
                "$ref": "#/components/schemas/CylinderShape"
              },
              {
                "$ref": "#/components/schemas/GridShape"
              }
            ],
            "title": "Shape"
          }
        },
        "required": [
          "shape",
          "ledOrder",
          "confirmed"
        ],
        "title": "Placement",
        "type": "object"
      },
      "PlacementIn": {
        "description": "PUT /lights/{id}/placement (ruling 16). Without ledOrder, a shape of the same kind\nkeeps its order and a new kind takes its first.",
        "properties": {
          "ledOrder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ledorder"
          },
          "shape": {
            "discriminator": {
              "mapping": {
                "bent-line": "#/components/schemas/BentLineShape",
                "cylinder": "#/components/schemas/CylinderShape",
                "grid": "#/components/schemas/GridShape",
                "line": "#/components/schemas/LineShape",
                "point": "#/components/schemas/PointShape"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PointShape"
              },
              {
                "$ref": "#/components/schemas/LineShape"
              },
              {
                "$ref": "#/components/schemas/BentLineShape"
              },
              {
                "$ref": "#/components/schemas/CylinderShape"
              },
              {
                "$ref": "#/components/schemas/GridShape"
              }
            ],
            "title": "Shape"
          }
        },
        "required": [
          "shape"
        ],
        "title": "PlacementIn",
        "type": "object"
      },
...
      "SubZone": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "polygon": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Polygon",
            "type": "array"
          },
          "room": {
            "title": "Room",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "room",
          "polygon"
        ],
        "title": "SubZone",
        "type": "object"
      },
      "SubZoneIn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "polygon": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Polygon",
            "type": "array"
          },
          "room": {
            "title": "Room",
            "type": "string"
          }
        },
        "required": [
          "name",
          "room",
          "polygon"
        ],
        "title": "SubZoneIn",
        "type": "object"
      },
      "SubZoneUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "polygon": {
            "anyOf": [
              {
                "items": {
                  "maxItems": 2,
                  "minItems": 2,
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon"
          },
          "room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          }
        },
        "title": "SubZoneUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "dj-ledfx",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/home": {
      "get": {
        "operationId": "get_home_api_home_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Home"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Home"
      },
      "put": {
        "description": "North, ceiling, beams and location; only what's sent changes.",
        "operationId": "update_home_api_home_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Home"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Home"
      }
    },
    "/api/home/anchors": {
      "post": {
        "operationId": "add_anchor_api_home_anchors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnchorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anchor"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Anchor"
      }
    },
    "/api/home/anchors/{anchor_id}": {
      "delete": {
        "operationId": "delete_anchor_api_home_anchors__anchor_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "anchor_id",
            "required": true,
            "schema": {
              "title": "Anchor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Anchor"
      },
      "put": {
        "operationId": "update_anchor_api_home_anchors__anchor_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "anchor_id",
            "required": true,
            "schema": {
              "title": "Anchor Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnchorUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anchor"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Anchor"
      }
    },
    "/api/home/subzones": {
      "post": {
        "operationId": "add_sub_zone_api_home_subzones_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubZoneIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubZone"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Sub Zone"
      }
    },
    "/api/home/subzones/{sub_zone_id}": {
      "delete": {
        "description": "A running sub-zone is turned off and its lights put back (Review Focus 4).",
        "operationId": "delete_sub_zone_api_home_subzones__sub_zone_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sub_zone_id",
            "required": true,
            "schema": {
              "title": "Sub Zone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Sub Zone"
      },
      "put": {
        "operationId": "update_sub_zone_api_home_subzones__sub_zone_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "sub_zone_id",
            "required": true,
            "schema": {
              "title": "Sub Zone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubZoneUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubZone"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Sub Zone"
      }
    },
...
    "/api/lights/{light_id}/placement": {
      "delete": {
        "description": "Take a light off the map (§8.6's \"Remove from the map\").",
        "operationId": "remove_placement_api_lights__light_id__placement_delete",
        "parameters": [
          {
            "in": "path",
            "name": "light_id",
            "required": true,
            "schema": {
              "title": "Light Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Placement"
      },
      "put": {
        "description": "Place or move a light, or a PC part by its device id (ruling 4). Moving a light\ndoesn't confirm it (spec §6.2).",
        "operationId": "place_light_api_lights__light_id__placement_put",
        "parameters": [
          {
            "in": "path",
            "name": "light_id",
            "required": true,
            "schema": {
              "title": "Light Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlacementIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Placement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Place Light"
      }
    },
    "/api/lights/{light_id}/placement/confirm": {
      "post": {
        "operationId": "confirm_placement_api_lights__light_id__placement_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "light_id",
            "required": true,
            "schema": {
              "title": "Light Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Placement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Placement"
      }
    },
...
  }
}